feat(cpp): out() rejects a float value -- the OUT port is integer-only
out(port, x) emitted gen_expr(x) + OUT, so a float x shipped its raw bit
pattern to the integer output port (meaningless). cppcomp has no fout and
the OUT instruction is integer-only, so reject a float value at compile time
with a message pointing at the cast: out(port, (int)x). Every existing test
already casts (out(0, (int)(h_hat[j]*GAIN_OUT+0.5f)), out(0, (int)sqrt(...)),
...), so all 51 still compile. Demo proc_cpp.cpp updated to cast (prod stays
visible as 3.75 via its GTKWave mirror; the OUT ships (int)prod = 3).

+++ Compilers/CPPComp/Tests/proc_cpp/Software/proc_cpp.cpp
@@ -22,6 +22,7 @@ void main(void)
     int   sum  = a + g_count;   // main sum  1   -> 10
     float prod = b * g_gain;    // main prod 2   -> 3.75
 
-    out(0, sum);     // 10
-    out(0, prod);    // 3.75
+    out(0, sum);          // 10
+    out(0, (int)prod);    // 3  (out() sends an integer word; cast the float --
+                          //     prod itself still shows as 3.75 in GTKWave)
 }
